refactor: コメントの追記とリファクタリング
