+++ packages/react/src/components/DropdownSelector/DropdownPopover.tsx
@@ -1,7 +1,23 @@
 import React, { useEffect, useRef } from 'react'
 import Popover, { PopoverProps } from './Popover'
 
-export function DropdownPopover({ children, state, ...props }: PopoverProps) {
+type DropdownPopoverProps = PopoverProps & {
+  value?: React.Key
+}
+
+/**
+ * DropdownSelectorの選択肢をを表示するためのPopover
+ * triggerRefの要素と同じ幅になる
+ * 表示の際にvalueが等しいDropdownMenuItemを中央に表示する
+ *
+ * @param param0
+ * @returns
+ */
+export function DropdownPopover({
+  children,
+  state,
+  ...props
+}: DropdownPopoverProps) {
   const ref = useRef<HTMLDivElement>(null)
   useEffect(() => {
     if (ref.current && props.triggerRef.current) {
